Make interval a u64 in config
This allows us to configure the checking interval in seconds instead of having
to use the `seconds` and `nanos` needed by the `Duration` type.
We are not using milliseconds or a float because we would never need to go
below one second of interval (going faster would be kind of spammy).