core,client: Remove some undoable TODO's
Since the Lagged message is sent from the websocket, it would be a bit
hacky to send a new initial status (even if it is doable by cloning the
app state).
An alternative to doing this would be to make clients always open a new
websocket when they get a `Lagged` message.